Allocate memory based on p_f_max
This is critical for larger systems
that will otherwise consume huge chunks of RAM
for 2e integrals.

We could do even better cause we do not need most
of the 2e integrals so we could store the sparse 2e matrix
in a more clever way, for example by reindexing virtual orbitals
and working only with (n_ov+1) x (n_ov+1) matrix for each
diagonalization step.